Refactor: Separate UserProfile component - UserProfile component의 길이가 너무 길어짐에 따라 Avatar, Nickname, Language로 분리하여 관리
Issue #519
